+++ Common/src/main/java/fuzs/resourcepackoverrides/client/data/PackSelectionOverride.java
@@ -1,6 +1,14 @@
 package fuzs.resourcepackoverrides.client.data;
 
-public record PackSelectionOverride(boolean forceCompatible, boolean fixedPosition, boolean required, boolean hidden) {
+import net.minecraft.network.chat.Component;
+import net.minecraft.server.packs.repository.Pack;
+import net.minecraft.server.packs.repository.PackCompatibility;
+import org.jetbrains.annotations.Nullable;
 
-    public static final PackSelectionOverride EMPTY = new PackSelectionOverride(false, false, false, false);
+public record PackSelectionOverride(@Nullable Component title, @Nullable Component description,
+                                    @Nullable Pack.Position defaultPosition, @Nullable PackCompatibility compatibility,
+                                    @Nullable Boolean fixedPosition, @Nullable Boolean required,
+                                    @Nullable Boolean hidden) {
+
+    public static final PackSelectionOverride EMPTY = new PackSelectionOverride(null, null, null, null, null, null, null);
 }