+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.0.2-1.19.2] - 2023-03-02
+### Added
+- Additional resource pack attributes for overriding: `title`, `description`, `default_position`
+### Changed
+- Some attributes are no longer ignored when set to `false`
+
 ## [v4.0.1-1.19.2] - 2023-02-28
 ### Fixed
 - Fixed a minor issue with loading a specified list of default resource packs